requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
chardet==4.0.0
discord.py==1.6.0
idna==3.1
idna-ssl==1.1.0
multidict==5.1.0
pkg-resources==0.0.0
python-dotenv==0.16.0
selenium==3.141.0
typing-extensions==3.7.4.3
urllib3==1.26.4
yarl==1.6.3
